@@ -1,48 +1,27 @@
 source 'https://rubygems.org'
 
+ruby '2.5.1'
 
-gem 'rails', '4.2.8'
+gem 'rails', '5.2.0'
 gem 'unicorn'
 gem 'pg'
+gem 'bootsnap', require: false
 
-# Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
+gem 'font-awesome-rails'
+gem 'sprockets', '>= 3.0.0'
+gem 'sprockets-es6'
 gem 'bourbon'
 
 gem 'stockfish'
 
-# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
-# gem 'turbolinks'
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-# gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-# gem 'sdoc', '~> 0.4.0', group: :doc
-
-gem 'font-awesome-rails'
-
-gem 'sprockets', '>= 3.0.0'
-gem 'sprockets-es6'
-
-group :development, :test do
-  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug'
-  gem 'quiet_assets'
-end
-
 group :development do
-  # Access an IRB console on exception pages or by using <%= console %> in views
-  gem 'web-console', '~> 2.0'
-
-  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-  gem 'spring'
+  gem 'web-console', '>= 3.3.0'
+  gem 'pry'
 end
-
