@@ -1,13 +1,12 @@
 class AnalysisController < ApplicationController
-  before_filter :set_analysis_options, :only => [:create]
+  before_action :set_analysis_options, only: [:create]
 
   def index
   end
 
   # Endpoint for requesting analysis for a position
-  #
   def create
-    render :json => StockfishAnalysis.new(params[:fen], @options).to_h.to_json
+    render json: StockfishAnalysis.new(params[:fen], @options).to_h.to_json
   end
 
   private
@@ -21,5 +20,4 @@ class AnalysisController < ApplicationController
       @options[:depth] = depth if (depth == 10 || depth == 16)
     end
   end
-
 end
